+++ changelog.md
@@ -6,6 +6,8 @@
 - Added classNamePrefix prop to JssProvider for better multi-tree support (#133).
 - Allowed nesting of JssProvider (#157).
 - Allow JssProvider rerender when used deep in the tree (#128).
+- Added `inject` option to `injectSheet` which allows you to filter props which are going to be passed to the child props (#130).
+- Removed the `sheet` prop. Now in order to get it back you need to specify the options `{inject: ['sheet']}`. Most of the time you should just use `props.classes`, which is provided by default.
 
 ## 7.2.0 / 2017-09-23